time double
Bench: 2505053

+++ src/timemanager.h
@@ -5,4 +5,4 @@
 #include <cmath>
 #include <cstdint>
 
-Time optimumTime(int64_t avaiableTime, int inc, int ply, int mtg);
+Time optimumTime(int64_t avaiableTime, int inc, int ply, double mtg);
